Fix erase mode for fill tool
This PR does 2 things:
1. Fixes the Erase mode button.
2. Fixes logic behind how the erase mode works.

Before the change, erase mode switch modified the selected color in drawing service, which had unwanted side effects:
* the selected color in color picker changed (and stayed changed even when the fill tool is deactivated)
* if the color is changed by user, erase mode broke, and then after erase mode is disabled, the color selected by user was reset to the active color at the moment when erase mode was enabled

To prevent this and make working of the tool consistent, I added a new drawing mode FillErase, which simplified things a lot. I also moved all the logic (which is not much) from the settings view to the fill tool, so that all the logic would be in the same place.

At this point the conditional statements in DrawingLayerNode about DrawingState ask for refactoring but I suggest we do it later, because I still don't fully understand how it works...